requirements.txt
annotated-types
anthropic
anyio
certifi
charset-normalizer
click
colorama
distro
exceptiongroup
fastapi
filelock
fsspec
h11
httpcore
httpx
huggingface-hub
idna
packaging
pydantic
pydantic-core
pyyaml
requests
setuptools
sniffio
starlette
tokenizers
tqdm
typing-extensions
urllib3
uvicorn
modal
python-dotenv
'r2r[parsing,eval]'